feat(layout): Make layout use server and add drafts to sidebar
