+++ package.json
@@ -54,6 +54,7 @@
     "@trpc/next": "^10.45.1",
     "@trpc/react-query": "^10.45.1",
     "@trpc/server": "^10.45.1",
+    "ai": "^2.2.36",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.0",
     "cmdk": "^0.2.1",
@@ -62,6 +63,7 @@
     "next": "^14.1.0",
     "next-auth": "^4.24.6",
     "next-themes": "^0.2.1",
+    "openai": "^4.28.0",
     "react": "18.2.0",
     "react-day-picker": "^8.10.0",
     "react-dom": "18.2.0",
