+++ src/app/page.tsx
@@ -3,6 +3,7 @@
 import { redirect } from 'next/navigation';
 import { SignInButton } from '@/app/_components/sign-in-button';
 import { getServerAuthSession } from '@/server/auth';
+import { Background } from '@/components/ui/background';
 
 export default async function Login() {
   const session = await getServerAuthSession();
@@ -15,11 +16,6 @@ export default async function Login() {
     <main className="h-full min-h-screen bg-gray-50">
       <div className="flex min-h-screen flex-1  flex-col items-center  justify-center py-12 sm:px-6 lg:px-8">
         <div className="sm:mx-auto sm:w-full sm:max-w-md">
-          <img
-            className="mx-auto h-10 w-auto"
-            src="https://tailwindui.com/img/logos/mark.svg?color=indigo&shade=600"
-            alt="Your Company"
-          />
           <h2 className="mt-6 text-center text-2xl font-bold leading-9 tracking-tight text-gray-900">
             Sign in to your Octo GPT account
           </h2>
@@ -27,6 +23,7 @@ export default async function Login() {
 
         <div className="mt-10 sm:mx-auto sm:w-full sm:max-w-[480px]">
           <div className="bg-white px-6 py-12 shadow sm:rounded-lg sm:px-12">
+            <Background className="mb-6" />
             <div>
               <div className="flex">
                 <SignInButton />
