+++ src/components/ui/background.tsx
@@ -1,14 +1,12 @@
-"use client";
+'use client';
 
-import Spline from "@splinetool/react-spline";
+import { cn } from '@/lib/utils';
+import Spline from '@splinetool/react-spline';
 
-export function Background() {
+export function Background({ className }: { className?: string }) {
   return (
-    <div className="background-object mb--100 -z-100 height-100 left-0">
-      <Spline
-        className="spline"
-        scene="https://prod.spline.design/2OJLJwr1DEbd6Bb4/scene.splinecode"
-      />
+    <div className={cn('h-44', className)}>
+      <Spline className="spline" scene="https://prod.spline.design/2OJLJwr1DEbd6Bb4/scene.splinecode" />
     </div>
   );
 }
